Add an express route to call youtube API on the backend instead of front end, refactor doubleClickHandler in builder.js to fetch from the backend #78
